Now folder is created using input file name
